// src/business/Interfaces/cricketMatchData.ts
 export interface CricketMatchData {
    ID: number;
    innings: number;
    overs: number;
    ballnumber: number;
    batter: string;
    bowler: string;
    non_striker: string;
    extra_type: string;
    batsman_run: number;
    extras_run: number;
    total_run: number;
    non_boundary: number;
    isWicketDelivery: number;
    player_out: string;
    kind: string;
    fielders_involved: string;
    BattingTeam: string;
}